make libp2p use from create function
